64-bit transition part 7: libcmd builtins
This is the seventh of the thickfold patch series, which enables ksh93
to operate within a 64-bit address space.

The parts of ksh93 affected by this commit are:
- The entirety of the libcmd builtins.

Remarks:
- Like the previous problematic getfsstat, sethostname also
  has a platform-dependent second argument that can be
  either signed or unsigned. This commit leaves the usage
  of that function (which passes a size_t value) as is.
- Moved the include directives in chmod and date to
  silence overlength string warnings.
- Removed unused macros from the fds code.
- Marked unused function parameters in grep to silence
  additional warnings.
- Use POSIX tcflag_t, speed_t and cc_t to silence
  warnings in stty.

Change in the number of warnings on Linux when compiling with clang using
-Wsign-compare -Wshorten-64-to-32 -Wsign-conversion -Wimplicit-int-conversion:
1,578 => 1,189 => 37 (progression from part 6 => part 7 => part 13)

Progresses https://github.com/ksh93/ksh/issues/592

+++ src/lib/libcmd/mktemp.c
@@ -66,7 +66,7 @@ b_mktemp(int argc, char** argv, Shbltin_t* context)
 	mode_t		mode = 0;
 	mode_t		mask;
 	int		fd;
-	int		i;
+	size_t		i;
 	int		quiet = 0;
 	int		unsafe = 0;
 	int*		fdp = &fd;
@@ -136,7 +136,7 @@ b_mktemp(int argc, char** argv, Shbltin_t* context)
 	}
 	if (t = strrchr(pfx, '/'))
 	{
-		i = ++t - pfx;
+		i = (size_t)(++t - pfx);
 		dir = fmtbuf(i);
 		memcpy(dir, pfx, i);
 		dir[i] = 0;
